feat: validate deletion from supply

+++ src/main/java/org/sistema/arroz/riceservice/modules/supplies/adapter/port/in/web/DeleteSupplyController.java
@@ -3,6 +3,8 @@ package org.sistema.arroz.riceservice.modules.supplies.adapter.port.in.web;
 import lombok.RequiredArgsConstructor;
 import org.sistema.arroz.riceservice.hexagonal.WebAdapter;
 import org.sistema.arroz.riceservice.modules.supplies.application.port.in.DeleteSupplyUseCase;
+import org.sistema.arroz.riceservice.modules.supplies.application.port.in.ValidateSupplyDeletionUseCase;
+import org.sistema.arroz.riceservice.modules.supplies.domain.SupplyNotValidToDeleteException;
 import org.springframework.web.bind.annotation.*;
 
 @WebAdapter
@@ -12,9 +14,12 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping(value = "/supplies")
 public class DeleteSupplyController {
     private final DeleteSupplyUseCase deleteSupplyUseCase;
+    private final ValidateSupplyDeletionUseCase validateSupplyDeletionUseCase;
 
     @DeleteMapping(value = "/{supplyId}")
     public Long deleteSupply(@PathVariable Long supplyId){
+        var result = validateSupplyDeletionUseCase.validateSupplyDeletion(supplyId);
+        if (result != null) throw new SupplyNotValidToDeleteException(supplyId, result);
         return deleteSupplyUseCase.deleteSupply(supplyId);
     }
 }
